Adjusted Screen/Made new map
I made a new map. Might change it to a different tile set but same layout.

For the world scale I changed it to x: .355 y: .34. It used to be x: 1 y: 1 It's a close enough fit.

The sprites now jump super high, so that will need to be adjusted. Not sure what other mechanics are changed by the change in the scale.